# services/backend-ts/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TickerGarden V1 Read API",
    "version": "5.4.0",
    "description": "Non-custodial TypeScript Serverless read API for the routes consumed by the current frontend."
  },
...
  "paths": {
    "/v1/users/{address}/activity": {
      "get": {
        "operationId": "listUserActivity",
        "description": "Finalized address-referenced protocol events, newest first, from the configured indexing start. All canonical batches and stored receipt commitments must be complete. Roles are event references, not verified transaction initiators or trading volume. A changed history returns 409; restart pagination. No transaction submission. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{40}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 1024
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivityPage"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityError"
                }
              }
            }
          },
          "409": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{txHash}": {
      "get": {
        "operationId": "getTransactionStatus",
        "description": "Read-only transaction observations. Distinguishes unknown, pending, confirmed, finalized and reorged; execution success/revert is separate. Confirms against a bracketed RPC head with journal provenance and retained orphan history. Unknown does not prove non-submission. Conflicting, stale or missing observations return 503. No signing or transaction submission. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "txHash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{64}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/holders": {
      "get": {
        "operationId": "getGlobalHolderCounts",
        "description": "TypeScript backend. Incremental exact global and per-STOCK positive-address counts at a finalized shared coverage anchor. Adjusted counts exclude the union of known protocol addresses. Counts are addresses, not people or reward eligibility. Missing coverage returns 503. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalHolderCountsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/series": {
      "get": {
        "operationId": "getGlobalFlowSeries",
        "description": "TypeScript Serverless backend. Verified execution flows grouped by STOCK/Quote in aligned [from,to) buckets. Empty buckets retain zero-flow groups; no synthetic trades, prices, balances or historical market counts. Maximum 2000 buckets and 100000 market-bucket cells, in addition to complete-history reader budgets. Incomplete/oversized history returns 503. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "5m",
                "15m",
                "1h",
                "4h",
                "1d"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalFlowSeriesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/overview": {
      "get": {
        "operationId": "getGlobalStatistics",
        "description": "TypeScript Serverless backend. Market and registered STOCK counts at the projection checkpoint (including retired registrations), execution flows within [from,to). Groups remain separated by STOCK and Quote; zero assetUid with binding=unbound explicitly means no STOCK binding. No implicit USD, reserve or holder total. One verified snapshot; maximum 1000 markets, 1000 registrations and 100000 executions, 10000 per market; incomplete or oversized history returns 503. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GlobalStatisticsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{marketId}/holders": {
      "get": {
        "operationId": "listMarketHolders",
        "description": "TypeScript Serverless backend. Complete Meme Token Transfer replay from creation through the finalized projection checkpoint. Balances in ascending address order, including excluded protocol accounts with explicit flags. Counts are addresses, not users or reward eligibility. Keep limit unchanged; on 409 discard prior pages and restart. Each page currently rereads bounded complete history (at most one million blocks and 100000 receipts/logs, 64MiB range payload budget, 10-second deadline); incomplete or oversized history returns 503. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "marketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{64}$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketHoldersResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderError"
                }
              }
            }
          },
          "409": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolderError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/markets/{marketId}/trades": {
      "get": {
        "operationId": "listMarketTrades",
        "description": "TypeScript Serverless backend. Newest-first canonical executions with explicit amount basis and caller identity confidence. Conversion summaries do not create extra executions. Keep market/from/to/limit unchanged across pages. HTTP 409 requires discarding previous pages and restarting without cursor. Currently revalidates the complete bounded interval per page (10000 executions maximum; database-side full-history coverage verification is subject to the request timeout). These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "marketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{64}$"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 2048
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketTradesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeError"
                }
              }
            }
          },
          "409": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{marketId}/detail": {
      "get": {
        "operationId": "getTokenDetail",
        "description": "Display-only approved token detail analytics. Prefers cached scheduled Dune results and falls back per section to finalized indexer coverage. Null means unavailable, never zero. Amounts are decimal strings; price and cap are Quote denominated. Protocol-balance exclusions define circulating supply. No execution or settlement authority. New launches may return a confirmed creation-transaction seed from the database; confirmation marks this provisional scope. Subsequent history comes from the independent indexer.",
        "parameters": [
          {
            "name": "marketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{64}$"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1H",
                "12H",
                "1D"
              ]
            }
          },
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "activity"
              ]
            },
            "description": "Only finalized recent trades and fee totals; skips chart, volume and holders. Omit for full summary."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/markets/{marketId}/candles": {
      "get": {
        "operationId": "getMarketCandles",
        "description": "TypeScript Serverless backend display-only canonical executions, including separately reported internal conversions. Aligned half-open Unix-second range; no synthetic trades or carry-forward prices. Missing coverage or configuration returns 503. At most 2000 buckets and 10000 trades. Database-side full-history coverage verification is subject to the request timeout. Independent coverage checkpoint, not a published snapshot revision. These nine history/transaction observation routes share a limit of four active requests per API instance. Excess requests return 503 analytics_unavailable with Retry-After: 5, without queueing.",
        "parameters": [
          {
            "name": "marketId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-f]{64}$"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1m",
                "5m",
                "15m",
                "1h",
                "4h",
                "1d"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(0|[1-9][0-9]*)$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketCandlesResponse"
                }
              }
            }
          },
          "400": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "405": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          },
          "503": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/prices/references": {
      "get": {
        "operationId": "listDisplayPriceReferences",
        "description": "Cached display estimates only, never an execution price or creation authorization. Testnet stock references may use a verified on-chain pool spot combined with native USD. No query parameters. Stale/unavailable prices are null.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayPriceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Query parameters unsupported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayPriceError"
                }
              }
            }
          },
          "405": {
            "description": "Read only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisplayPriceError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivityError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_query",
              "method_not_allowed",
              "activity_unavailable",
              "activity_page_changed",
              "analytics_unavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message",
          "requestId"
        ]
      },
...
      "TransactionError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_query",
              "method_not_allowed",
              "transaction_unavailable",
              "analytics_unavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message",
          "requestId"
        ]
      },
...
      "CandleError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_query",
              "method_not_allowed",
              "analytics_unavailable"
            ]
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message",
          "requestId"
        ]
      },
...
    }
  }
}